requirement.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==6.7
colorama==0.4.1
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
docopt==0.6.2
feedparser==5.2.1
gitsome==0.8.4
idna==2.8
ply==3.11
prompt-toolkit==2.0.10
Pygments==2.5.2
pytz==2016.10
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.25.7
wcwidth==0.1.7
